requirements.txt
gradio>=4.0
opencv-python
numpy
Pillow
pdf2image
google-cloud-vision
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
